Add new washing machine sensors  + sensors def moved to washing_machine.py
Total Electricity Used,
Total Water Used,
Mean water consumption,
Estimated weight (current cycle)

Move washing machine sensors def in  washing_machine.py
